frontend: router: Extract getRoute to own file
